feat(card_pays) : intégration de l'entête des card
