+++ components/icons/nuage.vue
@@ -6,8 +6,6 @@ defineProps({
 
 <template>
   <svg
-    width="80"
-    height="56"
     viewBox="0 0 80 56"
     fill="none"
     xmlns="http://www.w3.org/2000/svg"
