+++ scss/foundations/_variables.scss
@@ -3,7 +3,7 @@
 //media queries
 $small: 600px;
 $medium: 1000px;
-$large: 1800px;
+$large: 1500px;
 
 //Typographies
